# Conflicts:
#	db.sqlite3

+++ apps/quiz/models.py
@@ -3,6 +3,7 @@ from django.db import models
 
 class Team(models.Model):
     name = models.CharField(max_length=100)
-    points = models.IntegerField()
+    points = models.FloatField()
+
     def __str__(self):
         return self.name
